feat: even more api folder restructuring
