+++ webex_assistant_skills_sdk/api/mindmeld/models/__init__.py
@@ -1,6 +1,8 @@
+from webex_assistant_skills_sdk.api.mindmeld.models.dialogue_rule import MindmeldDialogueRule
 from webex_assistant_skills_sdk.api.mindmeld.models.processed_query import ProcessedQuery 
 
 
 __all__ = [
+    MindmeldDialogueRule,
     ProcessedQuery,
 ]
