+++ webex_assistant_skills_sdk/api/mindmeld/services/__init__.py
@@ -1,0 +1,6 @@
+from webex_assistant_skills_sdk.api.mindmeld.services.dialogue_manager import MindmeldDialogueManager
+
+
+__all__ = [
+    MindmeldDialogueManager,
+]
