+++ webex_assistant_skills_sdk/api/mindmeld/services/dialogue_manager.py
@@ -1,4 +1,4 @@
-class MMDialogueManager(DialogueManager):
+class MindmeldDialogueManager(DialogueManager):
     def add_rule(
         self,
         *,
