+++ webex_assistant_skills_sdk/api/simple/models/__init__.py
@@ -1,4 +1,4 @@
-from webex_assistant_skills_sdk.api.simple.models.rule import SimpleDialogueRule
+from webex_assistant_skills_sdk.api.simple.models.dialogue_rule import SimpleDialogueRule
 
 
 __all___ = [
