+++ webex_assistant_skills_sdk/api/simple/models/dialogue_rule.py
@@ -1,7 +1,7 @@
 import re
 from typing import Optional
 
-from webex_assistant_skills_sdk.api.shared.dialogue import DialogueRule
+from webex_assistant_skills_sdk.api.shared.models import DialogueRule
 
 
 class SimpleDialogueRule(DialogueRule[str]):
